+++ CHANGELOG.md
@@ -102,6 +102,7 @@ SPDX-License-Identifier: LGPL-3.0-or-later
 - Add support to handle inhomogeneous Dirichlet boundary conditions (IDBC) [#362](https://github.com/ikarus-project/ikarus/pull/362).
 - Add displacement-pressure (u-p) element [#391](https://github.com/ikarus-project/ikarus/pull/391).
 - Add transformation from material to two-point tangent moduli tensor [#398](https://github.com/ikarus-project/ikarus/pull/398).
+- Add functionality to pass quadrature rules to various finite elements [#399](https://github.com/ikarus-project/ikarus/pull/399).
 
 ## Release v0.4 (Ganymede)